Add a page for pre-generated User tokens to access backend
Page currently responde with 2 users, 1 with full crud permissions on a workitem
and 1 with only read workitem access.

Button 'Run as' currently does nothing, as we have no other UI
to interact with.

Page is intended to be use in the future to easy switch access level during
development/testing.

The page is backed by a REST service that generates tokens based on hardcoded
user/permissions. The REST service will return UnAuthorized if server is not
started in Development mode by passing in -dev on command line.