Improved Makefile
I've introduced a `help` target and comments on what each phony target does.

The new `all` target does nothing but triggering the `build` target.

The new `test` target does nothing but calling `go test` and this can be
extended in the future to have all containers started that we need to
test against some database or alike.

Here's the output of `make help`:

```
Make file targets
------------------

- all
	 Triggers these targets
	  - build

- build
	 Builds the binaries for the server and the client

- help
	 Shows all the commands and their description

- deps
	 Downloads the Go dependencies for this project

- generate
	 Bootstraps and generates code (using goa)

- clean
	 Removes the client and server binary

- dev
	 Sets up a developer environment

- test
	 Runs tests on the compiled Go binaries
```