Build script for Windows (jenkins ready)
This change contains a make.bat file in jenkins/windows/ that
can be used to prepare a Windows machine to build and test
this project.

We need these tools to build on Windows:

 * cygwin
 * go
 * glide
 * git
 * mercurial

The make.bat file can install and configure those tools locally
without requiring any user interaction.
I decided to write this file to be able to have a Jenkins job
on a Windows machine that knows nothing about any of the required
tools. This way we can modify our requirement in code.

make.bat can be seen like a Makefile on Linux. There are many targets
that can be executed. Among those targets is one called "help" that
tells more about what is available exactly.

A jenkins job can be configured to run "make.bat jenkins". This
ensures all tools are there and that the program is built and tested.

make.bat is configured to build in an out-of-source directory.
Therefore we have to copy everything from the the project folder
into the build directory.