Re-use user:group of caller in jenkins automatically
Previously I determined the username and group ID of the user that
started the jenkins build script and created a user with the exact
same ID inside the container.

This is not needed due to the "-u" switch on the "docker run" command.

This change greatly simplifies the build on a linux machine.

The build now runs in source and no longer out-of source which
helps to archive artifacts the were generated inside the container.