Check project location if GOPATH contains a PATH list
I figured it would be the easiest and most portable way
to do path checking in Go with an abstraction to the underlying OS.

But by the time of writing this I'm not sure Go can transparently
convert a Cygwin path to a Windows path. :(

Anyway this change moved some files only needed for for make
in a separate "./.make" directory. The primary reason being
that the go program I wrote ("check-gopath") contains "main"
function and as such it had to live in a separate directory.

When running go test on unit or integration test a "coverage-unit.out"
or "coverage-integration.out" file is written respectively.