WorkItem Update should happen by ID passed in URL
The WorkItem to be updated should be referred by its ID in the URL as a api/workitem/:id and the code should not be reading it from the payload since its not an editable field. Adding to it, the version number also was removed from the payload. This changed required a change in the goa design. The make task autogenerates sources inside the app/ dir

Fixes #95 #47

Signed-off-by: Shoubhik <shoubhik@dhcp35-156.lab.eng.blr.redhat.com>