List workitems (#120)
Fixes #101

This change implements a very simple "query by example" functionality where you can add two parameters to the url "api/workitem"

- page: <start>,<length> implements paging. If only one value is given, it is taken as the length parameters
- filter: you can pass in a json object string. Only work items matching (with '=') the field values given in the example json will be selected. 
Example: 
    http://localhost:8080/api/workitem?filter={"system.owner": "tmaeder","Name": "First","Type": "1"}&page=1,2

This change has a query parser, which produces expression trees, which will in turn be compiled for use with Gorm. While the query system architecture is in place, only a small amount of expressions is implemented (for example, only "=", no other comparisons). The intention is to extend the system as needed.

* First cut of simple work item listing
* Extract work item storage to a repository struct
* Always start from original db
* Extracted transaction support and errors
* Cleanup, doc & minor fixes
* Support for paging parameters
* Conditional value rendering depending on json context or regular context
* Fix test, update doc
* Fixed some comments
* Added comment
* Prevent sql injection
* Remove unused function