test/providers and tests requiring resources
Instead of specifying when a test shall skip we now let each test
specify what resources it depends on. These "resources" are governed by
the test/providers package and for the moment are nothing more than
special environment variables.

Currently a test can specify if it requires a database
(`providers.Database`) or if it requires to be run as a unit test
(`providers.UnitTest`). The later might be not needed but it is here for
documentation reasons and to have a place where to extend a test if it
required more dependencies.