Source interesting parts of jenkins slave env (#226)
This change sources interesting parts of the jenkins slave environment to the worker machine. 

This is a preparation for setting up an integration of an external coverage service, that might need some environment variables to be present in order to determine the CI type. Codecov.io is an example of such a service.

See almighty/almighty-jobs@e718d3aa96c1b933f269b138d39eab4dc6a881e7 to see how the jenkins-env file was created.