Run tests even if no DB conn was setup in TestMain (#257)
My proposal is simple: Run tests `m.Run()` no matter what the state of the environment variable is: Each test is responsible to specify which resource is required.

This way we can have unit tests and integration tests in one package without modifying a lot of the code right now. There will be a time when we might have a better way to setup and tear down a database connection. @aslakknutsen and I talked about this already. But this is a short term solution for the problem I think.

Fixes #255