Add realtime example for WorkItemType payload.
How:
Generally we add examples for each attribute separate. But type of `fields` attribute is a hash madhe up of UserDefinedType `fieldDefinition` and hence adding separate example for fields was failing. After many attempts I found this solution that we can add `Whole Example` along with all Attributes. Plus point is that `ExtendedType` is optional and is not promoted for usage as of now hence that is skipped.