Avoid too early expansion of $PATH variable (#319)
The first version we had caused problems on Windows systems: `$(PATH)`. This was problematic due to paths with whitespaces in them.

In my second approach I did this: `$$(PATH)`. The double `$$` was used to avoid a too early expansion of the `PATH` variable. Sadly this was also wrong and cause troubles (see [this mail thread](https://www.redhat.com/archives/almighty-public/2016-September/msg00219.html))

This time, the early expansion is correct: `$$PATH`.