Expose auth token keypair & github keys as env variables
Why?
Sensitive information used in production namely github secrets, encryption keys needs to be removed from the code. Having done that, it should not impair the average developer environment. There should be room for supplying the information in the form of env variables/config data , and in case none of that is done, there should be defaults to fall back upon.

What?
Added viper configuration for the same along with the default keys.

Fixes #337