close broken DB object
As per this document: http://go-database-sql.org/accessing.html
"the sql.DB object is designed to be long-lived"
The gorm.DB wraps sql.DB, so the above statement is still valid.
Since the ALM code is attempting to create new DB object, it's better
to close every time it fails to establish connection.

Updated gorm version to address two issues:

- to fix goroutine leak
  https://github.com/jinzhu/gorm/pull/1190
  (fix goroutine leak while calling db.Ping() error)

- Fix opening connection without any arguments.
  https://github.com/jinzhu/gorm/pull/1210