Set creator value while creating WI.
Why:
We need to set correct value for WI creator field.

What:
Every WI has list of fields, out of which SystemCreator is to be filled by system only.
Logged in user's UUID in string format should go in as value of SystemCreator in corresponding fields slice.

How:
Added a middleware to fetch, locate and set uuid in the context. If not found (for non-secured APIs) it will set it as "empty string".

Added functions to set and get CurrentUser from goa.context.