Remove dot-imports of goa design and design/apidsl packages
To be clear upfront: I'm NOT removing the dot-imports simply because
there's some warning about good Go practices on imports!

I found that no editor could autocomplete or jump to the declaration of
any of the words in the modified files. Being sick of jumping to the
online reference of goa, I decided to take the bumpy road and live with
a DSL that looks a little bit less okay that with the dot-import.

The benefit is, that I get autocompletion and can jump to the definition
of all the GOA symbols now.

With migration to JSON API ahead of us we're in desperate need of
control over all files in the design folder.