Export error types defined in models
Why:
All the exported error types in models/errors.go depend on models.simpleError type which is non-exported and therefore the exported error types are not usable unless invoked from the same package or if the errors.go is copied over the other specific package.

What:
Added method to access the non-exported methods.

fixes #407