Implement: WorkItemType.IsTypeOrSubtypeOf(name string)
Needed to add a method to fetch the model representation of work item
from the DB.

Added "Related planner item" link type that allows to link any planner
item to another one or a subtype of planner item.

Create and Update actions of work item link will check if the links ends
match the types or subtypes of those specified in the link type (see
ValidateCorrectSourceAndTargetType).

Also I've done lots of cleanups.