Conversion functions take object not pointer
Previously there was a strange error that @aslakknutsen discovered:

The two default work item link categories both had the same name:
"user". But it was only in the response not in the database or model
layer. That is why I've decide to pass input parameters rather than
pointers.

The underlying problem was that in order to create JSONAPI responses we
have to use pointers to strings for example. Whenever the pointer is
nil, it will be omitted by the JSON encoder. The pointer that was passed
in to the conversion functions was re-used and so that responses were
overritten.