Implement GET /api/workitem/:ID/relationships/links
Re-modelled the work item link respository
Outsourced controller code for WIL methods to be able to re-use it in
controller for /api/workitems/:ID/relationships/links.

The work item link repository's methods (list, show, delete, save, etc.)
now accept an additional string pointer. That pointer can be used to
optionally pass in the work item ID in case you're hitting the
repository from the controller for
/api/workitems/:id/relationships/links instead of just
/api/workitemlinks.

If a the afore mentioned string is not nil, we check that the work item
exists and may do some checking. For instance the source WI must be the
same as the given work item ID string for the create or save action.

Added tests for /api/workitems/:id/relationships/link based on /workitemlinks