Drop all auto soft-delete triggers for WorkItemLinking
If a WIL is manually deleted, then recreated with same target/source/type, and a WIT is updated,
all WIL of that type is reset to have same state as the WIT and causing unique constraint problems.

A WIT can not be deleted, it will only be disabled from view to be created
A WI is only ever soft deleted, but a WIL to a deleted item should still be displayable.
A WITC delete is a process and can probably never be deleted without a large user question of; what do you want to do with these?
A WILT delete can never happen, or similar to above.