Stored Github User information should be updated #537
Update on previous commit: fullname and avatar URL from GitHub
are stored in the identity during the update.

Added some tests in `login/service_whitebox_test.go` to verify
that the identity is created or updated. The test removes the
existing identity and users if it exists.
Using the `gormsupport.DBTestSuite` required some changes on all
tests in the file.

Fixes #537

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>