Separate between Work Item Type Field Type String and Markup (#477)
Converted the 'system.description' element into a 'markup' complex
element with 2 nested fields: 'content' and 'markup'.
Added a migration step to convert the JSON 'fields' of the existing
work items to use the new description format (ie, the 'system.description'
element in the JSON document is an object instead of a string literal).

Added a step to upgrade the fulltext search support in Postgres
to index the 'system.description'->>'content' element when a work item
is created or updated.

Added some possible values as constants, which are used
when importing remote work items from GitHub (`Markdown`) and
JIRA (`JiraWiki`).

When a new Work Item is created, the markup type associated
with the description is 'PlainText' unless it was specified in the
incoming request.

Note:
Incoming request can submit a work item description in the 'legacy' format
(i.e., a string) or in the new format. If the markup-type is missing, the
default 'PlainText' will be assigned.
However, the outgoing response will retain the legacy format for now, until
the UI layer is ready to support the 'MarkupContent' format.

Fixes #477

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>