Render the description in HTML using Markdown #634
Added "system.description.rendered" and "system.description.markup"
constants to return the rendered version of the description and the
markup language in the response, as sibblingis of the
"system.description" element.

Added a `rendering` package in which a function:
- returns the description if the markup is 'PlainText'
- uses the Blackfriday library (https://github.com/russross/blackfriday)
 to render the HTML from the given input.
- return nil otherwise.
A companion function returns a boolean to indicate if a given markup language
 is supported or not

In the future, this function can be refactored to support other markup
languages, and even delegate the processing to external services if needed.

Fixes #634

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>