Refactored code after code review
Moved the validation from the JSON input convertion into the
repository. The current `FieldDefinition.ConvertToModel` already
does the validation, but was missing the case where the field is
an empty string.

As a consequence, empty and nil/missing `system.title` are allowed
in the REST API (error is detected in the lower layer)

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>