Rendering description using markdown (2)
Added a migration script to set the markup to
'PlainText' instead of 'plain' to match the constant in the code.

Removed the error thrown during conversion from db to model if the
markup was invalid, yet keep preventing creation of work item
with invalid markup type
Forcing to default markup type if data from DB was corrupted/invalid.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>