Add an endpoint to render markup content in HTML in 'preview' mode #635
Added an endpoint at /api/render that takes a JSON document in the
POST request and convert the 'content` value of the input doc
using the `markup` value.
Returns a JSON response containing the rendered value if the markup
is supported.
Returns a BadRequest if the markup is not supported.

Fixes #635

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>