WIP Makefile: check-go-format, format-go-code + CI check (#719)
* Makefile: check-go-format format-go-code + CI check

check-go-format:
  - Exists with an error if there are files whose formatting differs
    from gofmt's

format-go-code:
  - Formats any go file that differs from gofmt's style

From this point onwards, the CI will check that GO code is correctly
formatted and if it is not, it will print an error that lists the files
that need to be changed.

In order to satisfy this CI check you can run "make format-go-code" to
have your code in proper shape. Then add those changes made to your
files and commit them.

This commit for example is based on a state of master where not all code
is correctly formatted, so it will not pass CI. I will make a second
commit that fixes this. This commit and the CI run are meant to be seen
as a demonstration that it works.

Try "make help" to find out more about the new make targets.

Thanks to @hectorj2f for coming up with this idea.

* Ran make format-go-code