Return the HTML rendering of comments #744
Added an extra field in the comment response to hold
the HTML rendered value based on the `body` and `markup`
of the comment.
Also, the generated HTML is escaped, preventing script
injection.

Refactored significantly the `work-item-comments_test.go`
file to make it more readable. Using methods instead of pure
functions allows for the same names in different TestSuite
in the same package.

Fixes #744
Fixes #724

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>