Assign creator identity for remote work items (#756)
Added a `RemoteWorkItem` structure to hold the relevant
data retrieved from the remote work items, including the
assignees' login and profile URL which is needed to look-up
the identity (or create a new one).

Added a column in `account.Identity` struct to store
the URL of the profile (`profile_url').
Renamed `provider` to `providerType` to better map i
with the purpose of the struct field.
Added indexes on the `identities` table.

Removed `remoteworkitem/flatten_test.go` as the sole test
already exists in `remoteworkitem/flatten_whitebox_test.go`

Formatted some JSON test file for better clarity.
Refactored `remoteworkitem/remoteworkitem_test.go` for better
clarity of code and to remove dupications of tests for GitHub
vs JIRA remote items.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>