Auto-delete links on a WI when WI gets deleted
If we don't auto-delete the links, then we get a 404 when we query for
the links on the still existing end of a link. This is because we try to
resolve both ends of the link and the one that has been deleted
obviously cannot be found.

For now we auto-delete the links associated to a WI in the WI
controller explicitly. Later we might want to have some registration
mechanism where different systems can hook into. For now this is
overkill.

This relates to https://github.com/almighty/almighty-core/issues/768