Greatly simplify the handling of IDs in linking
UUIDs can be passed through most of the time and don't have to be
checked.

Also the "type" field in relations is not checked for WILs, WILTs, and
WILCs because it is defined as an enum in the design folder. So REST
calls will not get through that have something different in the "type"
field anyway.