Record work item revisions (#840)
When a work item is created, modified or deleted (softly),
a copy of the data is stored in a new `work_item_revisions` table along
with a timestamp, the type of operation and the author/modifier
of the operation. In case of a deletion, the work item version is the same
as the last one (ie, before the deletion) but none of the fields are stored
in the revision table.

At the database level, there is a new a foreign key between the modifier_id
and an identity to verify that the identity of the modifier exists.
This required few changes in the existing test to comply with this constraint.

Fixes #840

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>