Convert work item type (WIT)  to JSONAPI (#721)
Fixes #674 #916 #848

Before this change, the work item types (WIT) API was using a custom REST format. Now it uses JSONAPI. I've centralized all WIT related things in the `design/workitemtype.go` file.

Before a WIT was referenced to by its `name`. I've added a new `id` field to the database table of type UUID and that is how WITs are referenced from this point onwards. The `name` field still exists but has no more restrictions on it. I've also added an optional `description` field to describe a WIT.

All former `fields` have been placed under `Data.Attributes.Fields` in the JSONAPI. The `name` and description are accessible through `Data.Attributes.Name` and `Data.Attributes.Description`.

The filter and search APIs no longer use a WIT's name but its ID to filter and search.

All references to a WIT in the DB have been update and now use the ID field.

All WITs now can specify a CSS icon. (Fixes #916)

All WIT Fields now have a label and a description. (Fixed #848)

I've also fixed a lot of tests not directly related to this PR. For example the search tests did not pass so I ripped out the undo script stuff and replaced it with a test suite that cleans up after each tests.
Also in quite a few places I stumbled over odd order of checks were an error was check after checking for a value.

## TODOs

- [x] Code wise, I have replaced the names of system defined WITs (e.g. `"bug"`, `"feature"`, etc.) with fixed UUIDs that MUST NOT be changed over the course of time.
In [the migration](https://github.com/kwk/almighty-core/blob/9530d1787fa243b3f5ebba7838721aa5a37d018b/migration/sql-files/034-wit-to-use-uuid.sql) I generate new UUIDs for the existing WITs in the DB. This is fine but I need to replace the UUIDs with the ones defined in the code if it is a system defined WIT. I'm on it. 
(Fixed in 15a21ed) 

- [ ] ~~Add links to WIT responses. Will be done when refactoring WIT repo to decouple from `app.*` package.~~

## JSONAPI representation of a WIT

Here's how the former `bug` type looks like.

```json
{
  "data": {
    "attributes": {
      "icon": "fa-bug",
      "description": "Description for Bug",
      "fields": {
        "system.area": {
          "label": "foo",
          "description": "bar",
          "required": false,
          "type": {
            "kind": "area"
          }
        },
        "system.assignees": {
          "label": "foo",
          "description": "bar",
          "required": false,
          "type": {
            "componentType": "user",
            "kind": "list"
          }
        },
        "system.created_at": {
          "label": "foo",
          "description": "bar",
          "required": false,
          "type": {
            "kind": "instant"
          }
        },
        "system.creator": {
          "label": "foo",
          "description": "bar",
          "required": true,
          "type": {
            "kind": "user"
          }
        },
        "system.description": {
          "label": "foo",
          "description": "bar",
          "required": false,
          "type": {
            "kind": "markup"
          }
        },
        "system.iteration": {
          "label": "foo",
          "description": "bar",
          "required": false,
          "type": {
            "kind": "iteration"
          }
        },
        "system.remote_item_id": {
          "label": "foo",
          "description": "bar",
          "required": false,
          "type": {
            "kind": "string"
          }
        },
        "system.state": {
          "label": "foo",
          "description": "bar",
          "required": true,
          "type": {
            "baseType": "string",
            "kind": "enum",
            "values": [
              "new",
              "open",
              "in progress",
              "resolved",
              "closed"
            ]
          }
        },
        "system.title": {
          "label": "foo",
          "description": "bar",
          "required": true,
          "type": {
            "kind": "string"
          }
        }
      },
      "name": "Bug",
      "version": 0
    },
    "id": "26787039-b68f-4e28-8814-c2f93be1ef4e",
    "type": "workitemtypes"
  }
}


```