Record work item comments revisions (#914)
Added a `comment_revisions` table to retain all changes on comments,
along with a timestamp, the Id of the author identity and the type
of operation that was performed.
When the comment is deleted, the 'body' and 'markup' fields are 'Nil'.
The list of revisions is retrieved in its historical order.

Also renamed the `workitem.WorkItemRevision` to `workitem.Revision`
and removed the `WorkItem` part in the revision constants.

Fixes #914

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>