Fixing 'show user' endpoint
After replacing the identity ID with the user ID in the
JSON response, the 'show user' endpoint needed to be fixed
by retrieving the user directly using the given ID param,
instead of passing by the identity.

Also, fixed the `500 Internal Server Error` response with
a more appropriate `404 Not Found` response when the given
ID does not match with any existing user.

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>