Index on work item link types should include the space_id (#1164)
When created, a work item link type didn't know about any space and thus its name was only allowed to be used once per link category. Now, with spaces, the same unique index shall span the name, the category and the space.

This change is important in order to allow using the same link type in multiple spaces, which is coming once we have space templates.