unique gorm create hook name (#1170)
This change checks if a GORM `Create` hook has already been registered under the name `"mighti:record"`. If that is the case, it will register the same hook under a different unique name. 

When you want to use `cleaner.DeleteCreatedEntries` outside of the and `*_test.go` file you need this.

For example, I want to create a temporary space and some work item types as well as work item link types, when somebody creates a space template. This is for validation purposes that the space template can be created.
Once I've created all those entries I want to get rid of them and just store the space template but no longer keep the temporary objects around. `cleaner.DeleteCreatedEntries` comes in handy here but without this change it won't work when invoked from within tests because the tests already called the function in order to clean up after them.

Consider this pseudo-code example

```go
func foo() {
  // with and without this change, the following defered call will delete d,c,b, and a which is fine.
  defer cleaner.DeleteCreatedEntries()
  a := create()
  b := create()
  bar()

func bar() {
  // without this change, the following defered call will cause objects d,c,b, and a to be deleted.
  // with this change only d and c are deleted.
  defer cleaner.DeleteCreatedEntries()
  c := create()
  d := create()
}

foo()
```