Allow update of identity username only once (#1160)
- Enable updation of email #1109
- Enable updation of username. 
- Allow updation of username only once by using the registrationCompleted boolean attributed

Handling the uniqueness constraints in Keycloak:

Email/username needs to be unique in keycloak, else keycloak returns a 500. As part of this change, platform api returns a more meaningful HTTP 409 conflict. This needs to be used by the UI to interpret why the 'update user profile' api call has failed. If keycloak updation fails, then platform db will NOT be updated.

Fixes #1118 , #1109