More relaxed forward name comparison to find children
We need this for the space templates where one type can have different
parent types. For example in the Scrum template
(https://docs.google.com/spreadsheets/d/1iQ3MBitXy5zWAvxWzHQKtXuv5SWs_uUCQvrqngfm5K8/edit#gid=1502690501)
a Product Backlog Item can be the parent of a Task and a Task can also
be a parent of a Task.

In order to distinguish those two relations when looking at the link
creation drop down box, we need to provide more informative link types
like "parent of Product Backlog Item" and "parent of Task". That is why
we need to relax the search for the correct link type a bit. This might
all change when we have link categories but until then, this is needed.