Support HTTP '409 Conflict' Errors when updating a work item (#1247)
Replaced `400 Bad Request` response with `409 Conflict` in case of
version conflict.
Added a function to load a workitem, with an option to issue a
`SELECT ... FOR UPDATE` when the work item is to be updated in the
current transaction, so it can be locked at the DB level and thus
prevent any concurrent write until the tx ends.

Fixes #1247

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>