Use fixed IDs for system-defined link types and categories (#1295)
In order to update the link system later I first need to make sure that work item link types and categories that have been created by the system use a pre-defined ID.

As a side-effect this will allow us to remove various loading functions that didn't properly load a category or type by an ID but by a name.

You can now pass in a complete or partial model of a category or type to the repository and have it created for you in the database. Before you needed to pass a bunch of parameters.